+++ app/src/main/java/com/freshdigitable/udonroad/TimelineFragment.java
@@ -16,7 +16,9 @@ import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.freshdigitable.udonroad.TimelineAdapter.OnUserIconClickedListener;
 import com.freshdigitable.udonroad.databinding.FragmentTimelineBinding;
+import com.freshdigitable.udonroad.datastore.TimelineStore;
 import com.freshdigitable.udonroad.ffab.FlingableFAB;
 import com.freshdigitable.udonroad.ffab.FlingableFABHelper;
 import com.freshdigitable.udonroad.ffab.OnFlingAdapter;
@@ -25,8 +27,11 @@ import java.util.List;
 
 import javax.inject.Inject;
 
+import rx.Observable;
+import rx.Subscription;
 import rx.android.schedulers.AndroidSchedulers;
 import rx.functions.Action1;
+import rx.schedulers.Schedulers;
 import twitter4j.Paging;
 import twitter4j.Status;
 import twitter4j.TwitterException;
@@ -35,15 +40,54 @@ public class TimelineFragment extends Fragment {
   @SuppressWarnings("unused")
   private static final String TAG = TimelineFragment.class.getSimpleName();
   private FragmentTimelineBinding binding;
-  private final TimelineAdapter tlAdapter = new TimelineAdapter();
+  private TimelineAdapter tlAdapter;
   private LinearLayoutManager tlLayoutManager;
   @Inject
   TwitterApi twitterApi;
+  @Inject
+  TimelineStore timelineStore;
+  private Subscription insertEventSubscription;
+  private Subscription updateEventSubscription;
+  private Subscription deleteEventSubscription;
 
   @Override
   public void onAttach(Context context) {
     super.onAttach(context);
     InjectionUtil.getComponent(this).inject(this);
+    tlAdapter = new TimelineAdapter(timelineStore);
+    timelineStore.open(context, getStoreName());
+    timelineStore.clear();
+    insertEventSubscription = timelineStore.subscribeInsertEvent()
+        .subscribe(new Action1<Integer>() {
+          @Override
+          public void call(Integer position) {
+            tlAdapter.notifyItemInserted(position);
+          }
+        });
+    updateEventSubscription = timelineStore.subscribeUpdateEvent()
+        .subscribe(new Action1<Integer>() {
+          @Override
+          public void call(Integer position) {
+            tlAdapter.notifyItemChanged(position);
+          }
+        });
+    deleteEventSubscription = timelineStore.subscribeDeleteEvent()
+        .subscribe(new Action1<Integer>() {
+          @Override
+          public void call(Integer position) {
+            tlAdapter.notifyItemRemoved(position);
+          }
+        });
+  }
+
+  @Override
+  public void onDetach() {
+    insertEventSubscription.unsubscribe();
+    updateEventSubscription.unsubscribe();
+    deleteEventSubscription.unsubscribe();
+    timelineStore.clear();
+    timelineStore.close();
+    super.onDetach();
   }
 
   @Nullable
@@ -78,9 +122,8 @@ public class TimelineFragment extends Fragment {
       }
     });
 
-    final TimelineAdapter adapter = getTimelineAdapter();
     final FlingableFAB fab = fabHelper.getFab();
-    adapter.setOnSelectedTweetChangeListener(
+    tlAdapter.setOnSelectedTweetChangeListener(
         new TimelineAdapter.OnSelectedTweetChangeListener() {
           @Override
           public void onTweetSelected(long statusId) {
@@ -92,13 +135,14 @@ public class TimelineFragment extends Fragment {
             fab.hide();
           }
         });
-    adapter.setLastItemBoundListener(new TimelineAdapter.LastItemBoundListener() {
+    tlAdapter.setLastItemBoundListener(new TimelineAdapter.LastItemBoundListener() {
       @Override
       public void onLastItemBound(long statusId) {
         fetchTweet(new Paging(1, 20, 1, statusId - 1));
       }
     });
-    binding.timeline.setAdapter(adapter);
+    tlAdapter.setOnUserIconClickedListener(userIconClickedListener);
+    binding.timeline.setAdapter(tlAdapter);
   }
 
   private final RecyclerView.AdapterDataObserver itemInsertedObserver
@@ -150,22 +194,20 @@ public class TimelineFragment extends Fragment {
   @Override
   public void onStart() {
     super.onStart();
-    final TimelineAdapter adapter = getTimelineAdapter();
-    adapter.registerAdapterDataObserver(itemInsertedObserver);
-    adapter.registerAdapterDataObserver(createdAtObserver);
+    tlAdapter.registerAdapterDataObserver(itemInsertedObserver);
+    tlAdapter.registerAdapterDataObserver(createdAtObserver);
     fetchTweet();
     setupOnFlingListener();
   }
 
   public void setupOnFlingListener() {
-    final TimelineAdapter adapter = getTimelineAdapter();
     fabHelper.getFab().setOnFlingListener(new OnFlingAdapter() {
       @Override
       public void onFling(Direction direction) {
         if (!isTweetSelected()) {
           return;
         }
-        final long id = adapter.getSelectedTweetId();
+        final long id = tlAdapter.getSelectedTweetId();
         if (Direction.UP.equals(direction)) {
           fetchFavorite(id);
         } else if (Direction.RIGHT.equals(direction)) {
@@ -201,18 +243,17 @@ public class TimelineFragment extends Fragment {
   public void onStop() {
     Log.d(TAG, "onStop: ");
     super.onStop();
-    final TimelineAdapter adapter = getTimelineAdapter();
-    adapter.unregisterAdapterDataObserver(itemInsertedObserver);
-    adapter.unregisterAdapterDataObserver(createdAtObserver);
+    clearSelectedTweet();
+    tlAdapter.unregisterAdapterDataObserver(itemInsertedObserver);
+    tlAdapter.unregisterAdapterDataObserver(createdAtObserver);
     tearDownOnFlingListener();
   }
 
   @Override
   public void onDestroyView() {
-    final TimelineAdapter adapter = getTimelineAdapter();
-    adapter.setLastItemBoundListener(null);
-    adapter.setOnSelectedTweetChangeListener(null);
-    adapter.setOnUserIconClickedListener(null);
+    tlAdapter.setLastItemBoundListener(null);
+    tlAdapter.setOnSelectedTweetChangeListener(null);
+    tlAdapter.setOnUserIconClickedListener(null);
     binding.timeline.setOnTouchListener(null);
     binding.timeline.setAdapter(null);
     super.onDestroyView();
@@ -249,7 +290,7 @@ public class TimelineFragment extends Fragment {
 
   private boolean canScroll() {
     return isVisible()
-        && !getTimelineAdapter().isStatusViewSelected()
+        && !tlAdapter.isStatusViewSelected()
         && statusDetail == null
         && !stopScroll
         && !isScrolledByUser
@@ -257,11 +298,11 @@ public class TimelineFragment extends Fragment {
   }
 
   private void addNewStatuses(List<Status> statuses) {
-    getTimelineAdapter().addNewStatuses(statuses);
+    tlAdapter.addNewStatuses(statuses);
   }
 
   private void addStatusesAtLast(List<Status> statuses) {
-    getTimelineAdapter().addNewStatusesAtLast(statuses);
+    tlAdapter.addNewStatusesAtLast(statuses);
   }
 
   public void scrollToTop() {
@@ -278,19 +319,20 @@ public class TimelineFragment extends Fragment {
   }
 
   public void clearSelectedTweet() {
-    getTimelineAdapter().clearSelectedTweet();
+    tlAdapter.clearSelectedTweet();
   }
 
   public boolean isTweetSelected() {
-    return getTimelineAdapter().isStatusViewSelected();
+    return tlAdapter.isStatusViewSelected();
   }
 
-  public void setUserIconClickedListener(TimelineAdapter.OnUserIconClickedListener listener) {
-    getTimelineAdapter().setOnUserIconClickedListener(listener);
+  private OnUserIconClickedListener userIconClickedListener;
+
+  public void setUserIconClickedListener(OnUserIconClickedListener listener) {
+    this.userIconClickedListener = listener;
   }
 
   private void fetchRetweet(final long tweetId) {
-    final TimelineAdapter adapter = getTimelineAdapter();
     final View rootView = binding.getRoot();
     twitterApi.retweetStatus(tweetId)
         .observeOn(AndroidSchedulers.mainThread())
@@ -298,7 +340,7 @@ public class TimelineFragment extends Fragment {
             new Action1<Status>() {
               @Override
               public void call(Status status) {
-                adapter.addNewStatus(status);
+                tlAdapter.addNewStatus(status);
               }
             },
             new Action1<Throwable>() {
@@ -311,7 +353,6 @@ public class TimelineFragment extends Fragment {
   }
 
   private void fetchFavorite(final long tweetId) {
-    final TimelineAdapter adapter = getTimelineAdapter();
     final View rootView = binding.getRoot();
     twitterApi.createFavorite(tweetId)
         .observeOn(AndroidSchedulers.mainThread())
@@ -319,7 +360,7 @@ public class TimelineFragment extends Fragment {
             new Action1<Status>() {
               @Override
               public void call(Status status) {
-                adapter.addNewStatus(status);
+                tlAdapter.addNewStatus(status);
               }
             },
             new Action1<Throwable>() {
@@ -380,10 +421,6 @@ public class TimelineFragment extends Fragment {
     return twitterApi;
   }
 
-  protected TimelineAdapter getTimelineAdapter() {
-    return tlAdapter;
-  }
-
   @Override
   public void onHiddenChanged(boolean hidden) {
     super.onHiddenChanged(hidden);
@@ -412,5 +449,40 @@ public class TimelineFragment extends Fragment {
 
   public void setFABHelper(FlingableFABHelper flingableFABHelper) {
     this.fabHelper = flingableFABHelper;
+  }
+
+  public String getStoreName() {
+    return "home";
+  }
+
+  protected void fetchTweet(final Observable.OnSubscribe<List<Status>> onSubscribe) {
+    Observable.create(onSubscribe)
+        .subscribeOn(Schedulers.io())
+        .observeOn(AndroidSchedulers.mainThread())
+        .subscribe(
+            new Action1<List<Status>>() {
+              @Override
+              public void call(List<Status> statuses) {
+                timelineStore.upsert(statuses);
+              }
+            },
+            new Action1<Throwable>() {
+              @Override
+              public void call(Throwable throwable) {
+                Log.e(TAG, "fetch: ", throwable);
+              }
+            });
+  }
+
+  protected static <T extends Fragment> T getInstance(T fragment, long userId) {
+    final Bundle args = new Bundle();
+    args.putLong("user_id", userId);
+    fragment.setArguments(args);
+    return fragment;
+  }
+
+  protected long getUserId() {
+    final Bundle arguments = getArguments();
+    return arguments.getLong("user_id");
   }
 }
