+++ app/src/main/java/com/freshdigitable/udonroad/UserFavsFragment.java
@@ -16,7 +16,7 @@ import twitter4j.TwitterException;
 /**
  * Created by akihit on 2016/06/07.
  */
-public class UserFavsFragment extends RealmTimelineFragment {
+public class UserFavsFragment extends TimelineFragment {
   @Override
   public String getStoreName() {
     return "user_favs";
